+++ expfactory/database/sqlite.py
@@ -90,9 +90,9 @@ Base = declarative_base()
 
 
 def init_db(self):
-    """initialize the database, with the default database path or custom of
+    """
+    Initialize the database, with the default database path or custom of
     the format sqlite:////scif/data/expfactory.db
-
     """
 
     # Database Setup, use default if uri not provided